Module_B/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
bcrypt==4.1.2
mysql-connector-python==8.3.0
matplotlib==3.8.0
pandas==2.1.0
jupyter==1.0.0
